+++ apps/main/src/components/Header/index.jsx
@@ -10,7 +10,7 @@ const Header = () => {
 
   return (
     <>
-      <header className="sticky top-0 z-50 shadow-md bg-base-100 navbar flex items-center justify-between px-6 py-3 text-base-content border-base-300">
+      <header className="fixed top-0 z-50 shadow-md bg-base-100 navbar flex items-center justify-between px-6 py-3 text-base-content border-base-300">
         <Link to="/" className="flex items-center gap-2" aria-label="Trang chủ">
           <span className="font-semibold gradient-text disable-select">
             Locket Dio
