+++ apps/main/src/services/LocketDioServices/index.js
@@ -8,6 +8,4 @@ export * from "./PayloadServices";
 export * from "./ApplyCoupon";
 export * from "./ActionMoments";
 export * from "./ChatServices";
-export * from "./RollcallServices";
-export * from "./FetchUserServices"
 export * from "./RequestServices";
