+++ apps/main/src/services/index.js
@@ -1,3 +1,4 @@
 export * from "./ExtensionsServices";
 export * from "./LocketDioServices";
 export * from "./BrowserServices";
+export * from "./LocketServices";
